Prototype for both raml and canonical validation checked in. Have to check the pros and cons of both the methods.
Canonical validation: canonical transformation currently happens in the next flow after ACK sent to 3ps. Had to alter the flow in case if we decided to go with this approach.
The errors are canonical related and might not make any sense that this field (canonical field) is required etc., to the client.

Raml schema validation: Will fail when the 1'st validation fails. Will not give all the errors in 1 shot. Having this schema gives a clear picture of what's expected in a language neutral way for the client.